+++ src/apps/frontend/next.config.ts
@@ -1,7 +1,10 @@
 import type { NextConfig } from "next";
+import { logEnvVariables } from "../../../.scripts/log-env-variables.ts";
 
 
 const isMobile = process.env.NEXT_PUBLIC_PLATFORM === "mobile";
+
+logEnvVariables();
 
 const nextConfig: NextConfig = isMobile
     ? {